Plan on republishing the Firefox extension
The Mozilla add-on store folks were less than useful when they blocked the review of the last iteration of the extension and would not follow up on my questions regarding what was I supposed to change. I ran out of patience, let the review window lapse and they removed the extension like the absolute dinguses they are. Some time has passed and I feel like I am ready for another round of probably non-interactive reviews but maybe they won't be and the extension will actually be published again.